Acks across servers
Changed client so that it acks specific messages. Hopefully this will have servers check all of their messages destined for a specific client. and only delete the specific messages that the client acks, not all of the messages with the same seq number. This ack should be sent across servers so that every server will delete the specific messages being acked by a client.