Sharring client sockets.
We fixed the issue of having multiple users on the same Ip address by making the client code 'request' use of the socket by attempting to create it. We also implemented a buffer in user_listen() so that we can simply check before-hand if we are about to print a duplicate message to the user.